* Navigation CSS Modification: - Ensure the navigation menu remains fixed even when the scrollbar is scrolled down.
* Header CSS Modification:
- Fix the header to the top using the fixed property and add a blur effect.

* Consolidate states related to the test update into a single file.
