+++ apps/admin/components/layout/Navigation/Index.tsx
@@ -1,29 +1,36 @@
 'use client';
 
-import React, { ReactNode, useState } from 'react';
 import { Breadcrumb, Layout, Menu, theme } from 'antd';
+import cx from 'classnames';
+import React, { ReactNode, useEffect, useState } from 'react';
 
 import styles from './index.module.scss';
-import HeaderComponent from '../Header';
 import items from './Items';
+import HeaderComponent from '../Header';
 
 const { Header, Content, Footer, Sider } = Layout;
 
 const Navigation: React.FC<{ children: ReactNode }> = ({ children }) => {
   const [collapsed, setCollapsed] = useState(false);
+
   const {
     token: { colorBgContainer, borderRadiusLG },
   } = theme.useToken();
 
   return (
-    <Layout style={{ minHeight: '100vh' }}>
-      <Sider collapsible collapsed={collapsed} onCollapse={(value) => setCollapsed(value)}>
+    <Layout style={{ minHeight: '100vh' }} hasSider>
+      <Sider
+        collapsible
+        collapsed={collapsed}
+        style={{ overflow: 'auto', height: '100vh', position: 'fixed', left: 0, top: 0, bottom: 0, zIndex: 10 }}
+        onCollapse={(value) => setCollapsed(value)}
+      >
         <Menu theme="dark" mode="inline" items={items} />
       </Sider>
-      <Layout className="min_wrap_w_size">
+      <Layout className="min_wrap_w_size" style={{ marginLeft: collapsed ? 80 : 200 }}>
         <Header className={styles.headerWrap} style={{ background: borderRadiusLG }}>
           <div className={styles.logoWrap}>MongBit Admin</div>
-          <div className={styles.headerBox}>
+          <div className={cx(styles.headerBox)}>
             <HeaderComponent />
           </div>
         </Header>
