--- apps/admin/states/testImageState.ts
@@ -1,8 +1,0 @@
-import { atom } from 'recoil';
-
-export const initialMbtiImageArray = Array(17).fill(undefined)
-
-export const mbtiImageState = atom<File[]>({
-  key: 'mbtiImageState',
-  default: initialMbtiImageArray,
-});
