--- apps/admin/types/selectOptions.ts
@@ -1,6 +1,0 @@
-export type ContentsSelectOptions = {
-  value: string;
-  label: string;
-  page: number;
-  disabled?: boolean;
-};
